comprehensive-plan/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
pydantic==2.6.1
neo4j==5.14.1
spacy==3.7.2
langchain==0.1.13
langchain-community==0.0.30
langchain-google-genai==0.0.6
langgraph==0.0.26
python-dotenv==1.0.1
typer==0.9.0
rich==13.7.0
faiss-cpu==1.7.4
sentence-transformers==2.2.2
unstructured[all-docs]
PyPDF2==3.0.1
python-magic==0.4.27 